acmp 492 wrong answer 38

+++ c/acmp/acmp492.cpp
@@ -3,14 +3,14 @@
 using namespace std;
 
 int main(int argc, char **argv) {
-	int x0, y0;
+	int64_t x0, y0;
 	cin >> x0 >> y0;
-	int vx, vy;
+	int64_t vx, vy;
 	cin >> vx >> vy;
-	int v, t, d;
+	int64_t v, t, d;
 	cin >> v >> t >> d;
-	int x1 = x0+vx*t;
-	int y1 = y0+vy*t;
+	int64_t x1 = x0+vx*t;
+	int64_t y1 = y0+vy*t;
 	int64_t dsp2 = d+v*t;
 	dsp2 *= dsp2;
 	int64_t dsm2 = d-v*t;
